chore: prepare for test

+++ server/src/main/java/com/github/dtprj/dongting/raft/store/LogFileQueue.java
@@ -548,7 +548,7 @@ class LogFileQueue extends FileQueue implements FileOps {
 
     @Override
     public long filePos(long absolutePos) {
-        return (int) (absolutePos & fileLenMask);
+        return absolutePos & fileLenMask;
     }
 
     @Override
